+++ src/context/CartContext.tsx
@@ -1,4 +1,6 @@
-import React, { createContext, ReactNode, useState } from "react";
+import React, { createContext, ReactNode, useEffect, useState } from "react";
+import { getProducts } from "../services/api/products";
+import type { IProduct } from "../services/api/products";
 
 export type Cart = {
   productId: number | undefined;
@@ -8,6 +10,7 @@ export type Cart = {
 export interface ICartContext {
   cart: Cart;
   setCart: React.Dispatch<React.SetStateAction<Cart>>;
+  products: IProduct[]
 }
 
 
@@ -17,6 +20,7 @@ const defaultState = {
     productQuantity: undefined,
   },
   setCart: (cart: Cart) => {},
+  products: [],
 } as ICartContext;
 
 type CartProviderProps = {
@@ -27,10 +31,29 @@ type CartProviderProps = {
 export const CartContext = createContext(defaultState);
 
 export default function CartProvider({children}: CartProviderProps) {
+
   const [cart, setCart] = useState<Cart>(defaultState.cart);
+  const [products, setProducts] = useState<IProduct[]>([])
+
+  async function fetchProducts() {
+    try {
+      // setLoading(true)
+      const response = await getProducts()
+      setProducts(response.data)
+    } catch (e) {
+      console.log('Error', e)
+    } finally {
+      // setLoading(false)
+    }
+  }
+
+  useEffect(() => {
+    fetchProducts()
+  }, [])
+
 
   return (
-    <CartContext.Provider value={{cart, setCart}}>
+    <CartContext.Provider value={{cart, setCart, products}}>
       {children}
     </CartContext.Provider>
   )
